Replaces getServerSession with slimmer version
